Rework network communication

+++ operators/fbx_transfer.py
@@ -1,23 +1,22 @@
 import bpy
-import time
 
 from ..utils import file_handling
+from ..utils.server_socket import ServerSocket
 from ..utils.csc_handling import CascadeurHandler
-from . import reciever
 
 
 def import_fbx(file_path: str) -> list:
     # Set import settings
     bpy.ops.import_scene.fbx(
         filepath=file_path,
-        use_anim=True,  # Don't import animations
-        use_image_search=True,  # Try to locate missing images
-        force_connect_children=False,  # Don't parent objects to armature bones
-        automatic_bone_orientation=False,  # Don't automatically orient bones
-        use_prepost_rot=False,  # Don't apply pre/post rotation
-        ignore_leaf_bones=True,  # Ignore leaf bones (not parented)
-        primary_bone_axis="Y",  # Set the primary bone axis to Y
-        secondary_bone_axis="X",  # Set the secondary bone axis to X
+        use_anim=True,
+        use_image_search=True,
+        force_connect_children=False,
+        automatic_bone_orientation=False,
+        use_prepost_rot=False,
+        ignore_leaf_bones=True,
+        primary_bone_axis="Y",
+        secondary_bone_axis="X",
         global_scale=1.0,
         use_manual_orientation=False,
         axis_forward="-Z",
@@ -57,19 +56,32 @@ class CBB_OT_import_cascadeur_fbx(bpy.types.Operator):
     bl_idname = "cbb.import_cascadeur_fbx"
     bl_label = "Import Cascadeur Scene"
 
-    def execute(self, context):
-        CascadeurHandler().execute_csc_command("commands.externals.temp_exporter.py")
-        bpy.ops.cbb.start_server()
-        data = reciever.recieved_data
-        time.sleep(2)
-        if data:
-            file_handling.wait_for_file(data)
-            import_fbx(data)
-            file_handling.delete_file(data)
-            reciever.recieved_data = None
-        else:
+    server_socket = None
+
+    def modal(self, context, event):
+        if event.type == "ESC":
+            self.server_socket.close()
+            self.server_socket = None
             return {"CANCELLED"}
-        return {"FINISHED"}
+
+        self.server_socket.run()
+
+        if self.server_socket.client_socket:
+            data = self.server_socket.receive_message()
+            if data:
+                print(str(data))
+                file_handling.wait_for_file(data)
+                import_fbx(data)
+                file_handling.delete_file(data)
+                return {"FINISHED"}
+
+        return {"PASS_THROUGH"}
+
+    def execute(self, context):
+        self.server_socket = ServerSocket()
+        CascadeurHandler().execute_csc_command("commands.externals.temp_exporter.py")
+        context.window_manager.modal_handler_add(self)
+        return {"RUNNING_MODAL"}
 
 
 class CBB_OT_import_action_to_selected(bpy.types.Operator):
@@ -77,6 +89,8 @@ class CBB_OT_import_action_to_selected(bpy.types.Operator):
 
     bl_idname = "cbb.import_cascadeur_action"
     bl_label = "Import Cascadeur Action"
+
+    ao = None
 
     @classmethod
     def poll(cls, context):
@@ -86,25 +100,34 @@ class CBB_OT_import_action_to_selected(bpy.types.Operator):
             and context.active_object.type == "ARMATURE"
         )
 
-    def execute(self, context):
-        ao = bpy.context.active_object
-        CascadeurHandler().execute_csc_command("commands.externals.temp_exporter.py")
-        bpy.ops.cbb.start_server()
-        data = reciever.recieved_data
-        time.sleep(2)
-        if data:
-            file_handling.wait_for_file(data)
-            imported_objects = import_fbx(data)
-            file_handling.delete_file(data)
-            reciever.recieved_data = None
-        else:
-            print("No data recieved")
+    def modal(self, context, event):
+        if event.type == "ESC":
+            self.server_socket.close()
+            self.server_socket = None
             return {"CANCELLED"}
 
-        actions = get_actions_from_objects(imported_objects)
-        apply_action(ao, actions)
-        delete_objects(imported_objects)
-        return {"FINISHED"}
+        self.server_socket.run()
+
+        if self.server_socket.client_socket:
+            data = self.server_socket.receive_message()
+            if data:
+                print(str(data))
+                file_handling.wait_for_file(data)
+                imported_objects = import_fbx(data)
+                file_handling.delete_file(data)
+                actions = get_actions_from_objects(imported_objects)
+                apply_action(self.ao, actions)
+                delete_objects(imported_objects)
+                return {"FINISHED"}
+
+        return {"PASS_THROUGH"}
+
+    def execute(self, context):
+        self.ao = bpy.context.active_object
+        self.server_socket = ServerSocket()
+        CascadeurHandler().execute_csc_command("commands.externals.temp_exporter.py")
+        context.window_manager.modal_handler_add(self)
+        return {"RUNNING_MODAL"}
 
 
 # Should be moved to a different place
